+++ src/helpers/Manifest.php
@@ -34,6 +34,13 @@ class Manifest
 
     const DEVMODE_CACHE_DURATION = 1;
 
+    const SUPPRESS_ERRORS_FOR_MODULES = [
+        'styles.js',
+        'commons.js',
+        'vendors.js',
+        'vendors.css',
+    ];
+
     // Protected Static Properties
     // =========================================================================
 
@@ -228,11 +235,14 @@ EOT;
         if ($manifest !== null) {
             // Make sure it exists in the manifest
             if (empty($manifest[$moduleName])) {
-                self::reportError(Craft::t(
-                    'seomatic',
-                    'Module does not exist in the manifest: {moduleName}',
-                    ['moduleName' => $moduleName]
-                ), $soft);
+                // Don't report errors for any files in SUPPRESS_ERRORS_FOR_MODULES
+                if (!in_array($moduleName, self::SUPPRESS_ERRORS_FOR_MODULES)) {
+                    self::reportError(Craft::t(
+                        'seomatic',
+                        'Module does not exist in the manifest: {moduleName}',
+                        ['moduleName' => $moduleName]
+                    ), $soft);
+                }
 
                 return null;
             }