+++ src/seomatic-config/productmeta/JsonLdContainer.php
@@ -74,7 +74,7 @@ return [
                     'seller'        => [
                         'id' => '{seomatic.site.identity.genericUrl}#identity',
                     ],
-                    'availability'  => 'http://schema.org/{% if object.product.unlimitedStock or object.product.totalStock > 0 %}InStock{% else %}OutOfStock{% endif %}',
+                    'availability'  => 'http://schema.org/{% if object.product.hasUnlimitedStock or object.product.totalStock > 0 %}InStock{% else %}OutOfStock{% endif %}',
                 ],
             ],
         ],