+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # SEOmatic Changelog
 
+## 3.3.27 - 2020.12.18
+### Changed
+* Split out vendors & commons chunks
+
+### Fixed
+* Changed `unlimitedStock` to `hasUnlimitedStock` for the Product JSON-LD defaults
+* Ensure that the `runtime.js` is included in the Field Settings and Field
+
 ## 3.3.26 - 2020.12.08
 ### Changed
 * Moved the CSS/JS buildchain over to webpack 5